+++ Tasks/UniversalPackagesV1/package.json
@@ -1,6 +1,6 @@
 {
   "name": "universalpackages",
-  "version": "1.270.0",
+  "version": "1.274.0",
   "description": "Download or publish universal packages.",
   "main": "universalMain.js",
   "scripts": {
@@ -24,7 +24,7 @@
     "@types/uuid": "^8.3.0",
     "azure-pipelines-task-lib": "^5.2.4",
     "azure-pipelines-tasks-artifacts-common": "^2.270.0",
-    "azure-pipelines-tasks-packaging-common": "^3.270.0",
+    "azure-pipelines-tasks-packaging-common": "^3.273.1",
     "azure-pipelines-tasks-utility-common": "3.270.0"
   },
   "devDependencies": {
