+++ Tasks/UniversalPackagesV1/task.json
@@ -9,7 +9,7 @@
   "category": "Package",
   "version": {
     "Major": 1,
-    "Minor": 273,
+    "Minor": 274,
     "Patch": 0
   },
   "runsOn": [
