+++ _generated/UniversalPackagesV1_Wif/package.json
@@ -1,6 +1,6 @@
 {
   "name": "universalpackages",
-  "version": "1.270.0",
+  "version": "1.274.0",
   "description": "Download or publish universal packages.",
   "main": "universalMain.js",
   "scripts": {
